api/docservice/api.py
from django.conf.urls import url, include
from .models import Document
from rest_framework import routers, serializers, viewsets

# Serializers define the API representation.
class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = '__all__

# ViewSets define the view behavior.
class DocumentViewSet(viewsets.ModelViewSet):
    queryset = Document.objects.all()
    serializer_class = DocumentSerializer

# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'documents', UserViewSet)
